# openapi.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "TaskManager API",
    "description": "Backend API for the distributed task management system. Handles task CRUD, authentication, role-based authorization, statistics, and publishes domain events to RabbitMQ.",
    "version": "1.0.0",
    "contact": {
      "name": "Person 1 — Backend Developer"
    }
  },
  "servers": [
    {
      "url": "http://localhost:3001",
      "description": "Local development"
    }
  ],
  "tags": [
    { "name": "Auth", "description": "Authentication (register / login)" },
    { "name": "Tasks", "description": "Task CRUD and state changes" },
    { "name": "Statistics", "description": "Task analytics and metrics" },
    { "name": "Admin", "description": "Admin-only user management" },
    { "name": "Health", "description": "Health check" }
  ],
  "paths": {
    "/api/v1/health": {
      "get": {
        "tags": ["Health"],
        "summary": "Health check",
        "operationId": "getHealth",
        "responses": {
          "200": {
            "description": "Service is healthy",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "status": { "type": "string", "example": "healthy" },
                    "timestamp": { "type": "string", "format": "date-time" }
                  }
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/auth/register": {
      "post": {
        "tags": ["Auth"],
        "summary": "Register a new user",
        "operationId": "register",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": { "$ref": "#/components/schemas/RegisterDto" },
              "example": {
                "email": "john@example.com",
                "password": "SecureP@ss123",
                "firstName": "John",
                "lastName": "Doe"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "User created",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/UserDto" }
              }
            }
          },
          "400": {
            "description": "Validation error",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/ErrorResponse" }
              }
            }
          }
        }
      }
    },
    "/api/v1/auth/login": {
      "post": {
        "tags": ["Auth"],
        "summary": "Login and receive JWT token",
        "operationId": "login",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": { "$ref": "#/components/schemas/LoginDto" },
              "example": {
                "email": "john@example.com",
                "password": "SecureP@ss123"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Login successful",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/AuthResponseDto" }
              }
            }
          },
          "400": {
            "description": "Validation error",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/ErrorResponse" }
              }
            }
          },
          "401": {
            "description": "Invalid credentials",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/ErrorResponse" }
              }
            }
          }
        }
      }
    },
    "/api/v1/tasks": {
      "get": {
        "tags": ["Tasks"],
        "summary": "List tasks (paginated, filterable)",
        "operationId": "getTasks",
        "parameters": [
          { "name": "page", "in": "query", "schema": { "type": "integer", "default": 1 } },
          { "name": "pageSize", "in": "query", "schema": { "type": "integer", "default": 20 } },
          { "name": "status", "in": "query", "schema": { "$ref": "#/components/schemas/TaskItemStatus" } },
          { "name": "priority", "in": "query", "schema": { "$ref": "#/components/schemas/TaskPriority" } },
          { "name": "assignedToId", "in": "query", "schema": { "type": "string", "format": "uuid" } },
          { "name": "sortBy", "in": "query", "schema": { "type": "string", "example": "createdAt" } },
          { "name": "sortDescending", "in": "query", "schema": { "type": "boolean", "default": false } }
        ],
        "responses": {
          "200": {
            "description": "Paginated list of tasks",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/PagedResultTaskDto" }
              }
            }
          }
        }
      },
      "post": {
        "tags": ["Tasks"],
        "summary": "Create a new task",
        "operationId": "createTask",
        "security": [{ "bearerAuth": [] }],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": { "$ref": "#/components/schemas/CreateTaskDto" },
              "example": {
                "title": "Implement caching layer",
                "description": "Add Redis caching to the task queries",
                "priority": "High",
                "dueDate": "2026-04-01T00:00:00Z"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Task created",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/TaskDto" }
              }
            }
          },
          "400": {
            "description": "Validation error",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/ErrorResponse" }
              }
            }
          },
          "401": { "description": "Unauthorized" }
        }
      }
    },
    "/api/v1/tasks/{id}": {
      "get": {
        "tags": ["Tasks"],
        "summary": "Get task by ID",
        "operationId": "getTaskById",
        "parameters": [
          { "name": "id", "in": "path", "required": true, "schema": { "type": "string", "format": "uuid" } }
        ],
        "responses": {
          "200": {
            "description": "Task found",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/TaskDto" }
              }
            }
          },
          "404": {
            "description": "Task not found",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/ErrorResponse" }
              }
            }
          }
        }
      },
      "put": {
        "tags": ["Tasks"],
        "summary": "Update a task",
        "operationId": "updateTask",
        "security": [{ "bearerAuth": [] }],
        "parameters": [
          { "name": "id", "in": "path", "required": true, "schema": { "type": "string", "format": "uuid" } }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": { "$ref": "#/components/schemas/UpdateTaskDto" },
              "example": {
                "title": "Updated title",
                "description": "Updated description",
                "priority": "Critical",
                "dueDate": "2026-04-15T00:00:00Z"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Task updated",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/TaskDto" }
              }
            }
          },
          "400": {
            "description": "Validation error",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/ErrorResponse" }
              }
            }
          },
          "401": { "description": "Unauthorized" },
          "404": {
            "description": "Task not found",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/ErrorResponse" }
              }
            }
          }
        }
      },
      "delete": {
        "tags": ["Tasks"],
        "summary": "Delete a task (soft delete)",
        "operationId": "deleteTask",
        "security": [{ "bearerAuth": [] }],
        "parameters": [
          { "name": "id", "in": "path", "required": true, "schema": { "type": "string", "format": "uuid" } }
        ],
        "responses": {
          "204": { "description": "Task deleted" },
          "401": { "description": "Unauthorized" },
          "404": {
            "description": "Task not found",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/ErrorResponse" }
              }
            }
          }
        }
      }
    },
    "/api/v1/tasks/{id}/status": {
      "patch": {
        "tags": ["Tasks"],
        "summary": "Change task status",
        "operationId": "changeTaskStatus",
        "security": [{ "bearerAuth": [] }],
        "parameters": [
          { "name": "id", "in": "path", "required": true, "schema": { "type": "string", "format": "uuid" } }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": { "$ref": "#/components/schemas/ChangeStatusDto" },
              "example": { "status": "InProgress" }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Status changed",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/TaskDto" }
              }
            }
          },
          "400": {
            "description": "Invalid status transition",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/ErrorResponse" }
              }
            }
          },
          "401": { "description": "Unauthorized" }
        }
      }
    },
    "/api/v1/tasks/{id}/assign": {
      "patch": {
        "tags": ["Tasks"],
        "summary": "Assign task to a user",
        "operationId": "assignTask",
        "security": [{ "bearerAuth": [] }],
        "parameters": [
          { "name": "id", "in": "path", "required": true, "schema": { "type": "string", "format": "uuid" } }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": { "$ref": "#/components/schemas/AssignTaskDto" },
              "example": { "userId": "3fa85f64-5717-4562-b3fc-2c963f66afa6" }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Task assigned",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/TaskDto" }
              }
            }
          },
          "400": {
            "description": "Invalid assignment",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/ErrorResponse" }
              }
            }
          },
          "401": { "description": "Unauthorized" }
        }
      }
    },
    "/api/v1/tasks/statistics": {
      "get": {
        "tags": ["Statistics"],
        "summary": "Get overall task statistics",
        "operationId": "getStatistics",
        "security": [{ "bearerAuth": [] }],
        "responses": {
          "200": {
            "description": "Statistics",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/TaskStatisticsDto" }
              }
            }
          },
          "401": { "description": "Unauthorized" }
        }
      }
    },
    "/api/v1/tasks/statistics/by-user": {
      "get": {
        "tags": ["Statistics"],
        "summary": "Get per-user task statistics",
        "operationId": "getStatisticsByUser",
        "security": [{ "bearerAuth": [] }],
        "responses": {
          "200": {
            "description": "Per-user statistics",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": { "$ref": "#/components/schemas/UserStatisticsDto" }
                }
              }
            }
          },
          "401": { "description": "Unauthorized" }
        }
      }
    },
    "/api/v1/tasks/statistics/timeline": {
      "get": {
        "tags": ["Statistics"],
        "summary": "Get task creation/completion timeline",
        "operationId": "getTimeline",
        "security": [{ "bearerAuth": [] }],
        "parameters": [
          { "name": "period", "in": "query", "schema": { "type": "string", "default": "daily", "enum": ["daily", "weekly", "monthly"] } }
        ],
        "responses": {
          "200": {
            "description": "Timeline data",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": { "$ref": "#/components/schemas/TimelineStatisticsDto" }
                }
              }
            }
          },
          "401": { "description": "Unauthorized" }
        }
      }
    },
    "/api/v1/admin/users": {
      "get": {
        "tags": ["Admin"],
        "summary": "List all users (Admin only)",
        "operationId": "getUsers",
        "security": [{ "bearerAuth": [] }],
        "parameters": [
          { "name": "page", "in": "query", "schema": { "type": "integer", "default": 1 } },
          { "name": "pageSize", "in": "query", "schema": { "type": "integer", "default": 20 } }
        ],
        "responses": {
          "200": {
            "description": "Paginated list of users",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/PagedResultUserDto" }
              }
            }
          },
          "401": { "description": "Unauthorized" },
          "403": { "description": "Forbidden — Admin role required" }
        }
      }
    },
    "/api/v1/admin/users/{id}/role": {
      "patch": {
        "tags": ["Admin"],
        "summary": "Change a user's role (Admin only)",
        "operationId": "changeUserRole",
        "security": [{ "bearerAuth": [] }],
        "parameters": [
          { "name": "id", "in": "path", "required": true, "schema": { "type": "string", "format": "uuid" } }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": ["role"],
                "properties": {
                  "role": { "$ref": "#/components/schemas/UserRole" }
                }
              },
              "example": { "role": "Manager" }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Role changed",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/UserDto" }
              }
            }
          },
          "400": {
            "description": "Invalid role",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/ErrorResponse" }
              }
            }
          },
          "401": { "description": "Unauthorized" },
          "403": { "description": "Forbidden — Admin role required" },
          "404": {
            "description": "User not found",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/ErrorResponse" }
              }
            }
          }
        }
      }
    }
  },
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "description": "JWT token obtained from POST /api/v1/auth/login"
      }
    },
    "schemas": {
      "RegisterDto": {
        "type": "object",
        "required": ["email", "password", "firstName", "lastName"],
        "properties": {
          "email": { "type": "string", "format": "email" },
          "password": { "type": "string", "minLength": 6 },
          "firstName": { "type": "string" },
          "lastName": { "type": "string" }
        }
      },
      "LoginDto": {
        "type": "object",
        "required": ["email", "password"],
        "properties": {
          "email": { "type": "string", "format": "email" },
          "password": { "type": "string" }
        }
      },
      "AuthResponseDto": {
        "type": "object",
        "properties": {
          "token": { "type": "string" },
          "expiresAt": { "type": "string", "format": "date-time" },
          "user": { "$ref": "#/components/schemas/UserDto" }
        }
      },
      "UserDto": {
        "type": "object",
        "properties": {
          "id": { "type": "string", "format": "uuid" },
          "email": { "type": "string", "format": "email" },
          "firstName": { "type": "string" },
          "lastName": { "type": "string" },
          "role": { "$ref": "#/components/schemas/UserRole" }
        }
      },
      "TaskDto": {
        "type": "object",
        "properties": {
          "id": { "type": "string", "format": "uuid" },
          "title": { "type": "string" },
          "description": { "type": "string" },
          "status": { "$ref": "#/components/schemas/TaskItemStatus" },
          "priority": { "$ref": "#/components/schemas/TaskPriority" },
          "dueDate": { "type": "string", "format": "date-time", "nullable": true },
          "assignedToId": { "type": "string", "format": "uuid", "nullable": true },
          "createdById": { "type": "string", "format": "uuid" },
          "createdAt": { "type": "string", "format": "date-time" },
          "updatedAt": { "type": "string", "format": "date-time" }
        }
      },
      "CreateTaskDto": {
        "type": "object",
        "required": ["title", "description", "priority"],
        "properties": {
          "title": { "type": "string" },
          "description": { "type": "string" },
          "priority": { "$ref": "#/components/schemas/TaskPriority" },
          "dueDate": { "type": "string", "format": "date-time", "nullable": true }
        }
      },
      "UpdateTaskDto": {
        "type": "object",
        "required": ["title", "description", "priority"],
        "properties": {
          "title": { "type": "string" },
          "description": { "type": "string" },
          "priority": { "$ref": "#/components/schemas/TaskPriority" },
          "dueDate": { "type": "string", "format": "date-time", "nullable": true }
        }
      },
      "ChangeStatusDto": {
        "type": "object",
        "required": ["status"],
        "properties": {
          "status": { "$ref": "#/components/schemas/TaskItemStatus" }
        }
      },
      "AssignTaskDto": {
        "type": "object",
        "required": ["userId"],
        "properties": {
          "userId": { "type": "string", "format": "uuid" }
        }
      },
      "TaskStatisticsDto": {
        "type": "object",
        "properties": {
          "totalTasks": { "type": "integer" },
          "byStatus": {
            "type": "object",
            "additionalProperties": { "type": "integer" },
            "example": { "Pending": 5, "InProgress": 3, "Completed": 10, "Cancelled": 1 }
          },
          "byPriority": {
            "type": "object",
            "additionalProperties": { "type": "integer" },
            "example": { "Low": 4, "Medium": 6, "High": 5, "Critical": 2 }
          },
          "overdueCount": { "type": "integer" }
        }
      },
      "UserStatisticsDto": {
        "type": "object",
        "properties": {
          "userId": { "type": "string", "format": "uuid" },
          "name": { "type": "string" },
          "totalAssigned": { "type": "integer" },
          "completed": { "type": "integer" },
          "inProgress": { "type": "integer" },
          "completionRate": { "type": "number", "format": "double" }
        }
      },
      "TimelineStatisticsDto": {
        "type": "object",
        "properties": {
          "period": { "type": "string" },
          "created": { "type": "integer" },
          "completed": { "type": "integer" }
        }
      },
      "ErrorResponse": {
        "type": "object",
        "properties": {
          "error": { "type": "string" },
          "details": { "type": "object", "nullable": true }
        }
      },
      "PagedResultTaskDto": {
        "type": "object",
        "properties": {
          "data": {
            "type": "array",
            "items": { "$ref": "#/components/schemas/TaskDto" }
          },
          "total": { "type": "integer" },
          "page": { "type": "integer" },
          "pageSize": { "type": "integer" }
        }
      },
      "PagedResultUserDto": {
        "type": "object",
        "properties": {
          "data": {
            "type": "array",
            "items": { "$ref": "#/components/schemas/UserDto" }
          },
          "total": { "type": "integer" },
          "page": { "type": "integer" },
          "pageSize": { "type": "integer" }
        }
      },
      "TaskItemStatus": {
        "type": "string",
        "enum": ["Pending", "InProgress", "Completed", "Cancelled"]
      },
      "TaskPriority": {
        "type": "string",
        "enum": ["Low", "Medium", "High", "Critical"]
      },
      "UserRole": {
        "type": "string",
        "enum": ["User", "Manager", "Admin"]
      }
    }
  }
}